refactor(defer): share resolver-building logic between defer block and metadata async
`build_defer_resolver_expression` (used for `ɵɵdefer`'s resolver argument)
and `compile_component_metadata_async_resolver` (used for
`ɵsetClassMetadataAsync`) emitted byte-equivalent `() => [import('./x').then((m) => m.X)]`
shapes via two independent builders. Consolidate by:

- Moving `build_defer_per_component_deps` from `component/transform.rs` into
  `component/defer_resolver.rs` (the natural home — it's a resolver concern,
  and transform.rs already depends on defer_resolver, not vice versa).
- Reducing `build_defer_resolver_expression` to a thin wrapper that builds
  the deps slice and delegates to `compile_component_metadata_async_resolver`.
- Dropping the now-unused `build_import_then_expression` helper.

No emit changes — both call sites already produced identical output for
every input shape (default, aliased, plain named, namespace, type-only).
Net `-104` lines across the two files.
